+++ Source/Gapotchenko.FX.Linq/EnumerableExtensions.cs
@@ -1,0 +1,9 @@
+﻿namespace Gapotchenko.FX.Linq;
+
+/// <summary>
+/// Provides extension methods for <see cref="IEnumerable{T}"/>.
+/// </summary>
+[EditorBrowsable(EditorBrowsableState.Never)]
+public static partial class EnumerableExtensions
+{
+}